requirements/base.txt
django==5.0.7
django-jazzmin==3.0.0
django-modeltranslation==0.19.4
django-cors-headers==4.4.0
django-rosetta==0.10.0

djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

drf-yasg==1.21.7

pillow==10.4.0
python-environ==0.4.54
setuptools==71.0.3

# Database
psycopg2-binary==2.9.9
